Remove Scopes and Claims
The access token should be opaque, and the interface to user information
should be perused by token introspection via the identity service.
